server/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud and OAuth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-cloud-secret-manager>=2.16.0,<3.0.0

# HTTP client
httpx==0.25.2


# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.1.0,<3.0.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Session management
starlette==0.27.0
itsdangerous==2.2.0

# Plaid SDK
plaid-python==35.0.0

# Cryptography for token encryption
cryptography>=41.0.0,<42.0.0

# Azure Cosmos DB
azure-cosmos>=4.5.1,<5.0.0

azure-functions